Reminder Model

+++ src/test/java/ch/bfh/bti7081/s2019/green/persistence/util/IdUtil.java
@@ -2,12 +2,12 @@ package ch.bfh.bti7081.s2019.green.persistence.util;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public class IdUtil {
-    private static final Map<Class<?>, AtomicLong> idsByClass = new ConcurrentHashMap<>();
+    private static final Map<Class<?>, AtomicInteger> idsByClass = new ConcurrentHashMap<>();
 
-    public static Long next(Class<?> clazz){
-        return idsByClass.computeIfAbsent(clazz, c -> new AtomicLong()).incrementAndGet();
+    public static Integer next(Class<?> clazz){
+        return idsByClass.computeIfAbsent(clazz, c -> new AtomicInteger()).incrementAndGet();
     }
 }
